refactor: improve OpenAI service organization
- Split OpenAIService into smaller services:
  - ChatService: chat message handling
  - PromptService: system prompt generation
  - ResponseParser: JSON response parsing
- Improved code organization and maintainability
- Better separation of concerns
